Release 10.1.0

+++ app/workers/unlink_from_shared_collections_worker.rb
@@ -9,6 +9,7 @@ class UnlinkFromSharedCollectionsWorker
     (users + groups).each do |entity|
       objects.each do |object|
         shared = entity.current_shared_collection
+        next unless shared.present?
         shared_link = shared.link_collection_cards.with_record(object).first
         if shared_link.present? && !object.can_view?(entity)
           shared_link.destroy
